# kubernetes/swagger-config copy.yaml
openapi: 3.0.3
info:
  title: Messenger API
  version: 1.0.1
  description: API мессенджера
servers:
  - url: /api
    description: REST API сервера
  - url: ws://localhost:8083
    description: WebSocket сервер для звонков
security:
  - BearerAuth: []
...
paths:
...
  /v1/calls/user/{userId}:
    post:
      tags:
        - Звонки и Комнаты
      summary: Позвонить пользователю
      description: Создает приватную комнату и инициирует прямой звонок конкретному пользователю
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID пользователя, которому звоним
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum: [audio, video]
                  description: Тип звонка
                  example: "audio"
      responses:
        '201':
          description: Звонок успешно инициирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  data:
                    type: object
                    properties:
                      call:
                        $ref: '#/components/schemas/Call'
                      room:
                        $ref: '#/components/schemas/Room'
        '400':
          description: Неверные параметры запроса
        '409':
          description: Пользователь уже в звонке
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: "У пользователя уже есть активный звонок"
...
  /v1/calls/{callId}/accept:
    put:
      tags:
        - Звонки и Комнаты
      summary: Принять звонок
      description: Принимает входящий звонок и подключается к нему
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID звонка
      responses:
        '200':
          description: Звонок успешно принят
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/Call'
        '404':
          description: Звонок не найден
        '409':
          description: Невозможно принять звонок в текущем статусе
  /v1/calls/{callId}/reject:
    post:
      tags:
        - Звонки и Комнаты
      summary: Отклонить звонок
      description: Отклоняет входящий звонок
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID звонка
      responses:
        '200':
          description: Звонок успешно отклонен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/Call'
        '404':
          description: Звонок не найден
        '409':
          description: Невозможно отклонить звонок
  /v1/calls/{callId}/end:
    post:
      tags:
        - Звонки и Комнаты
      summary: Завершить звонок
      description: Завершает текущий звонок
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID звонка
      responses:
        '200':
          description: Звонок успешно завершен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/Call'
        '404':
          description: Звонок не найден
        '409':
          description: Звонок уже завершен
  /v1/calls/{callId}/signal:
    post:
      tags:
        - Звонки и Комнаты
      summary: Отправить WebRTC сигнал
      description: Отправляет сигнальное сообщение WebRTC другому участнику звонка
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID звонка
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalingMessage'
      responses:
        '200':
          description: Сигнал успешно отправлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
        '404':
          description: Звонок не найден
        '400':
          description: Неверный формат сигнала
  /v1/calls/{callId}/invite:
    post:
      tags:
        - Звонки и Комнаты
      summary: Сгенерировать приглашение в комнату звонка
      description: Создает новый код приглашения для присоединения к комнате звонка
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID звонка
      responses:
        '200':
          description: Приглашение успешно создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      invite_code:
                        type: string
                        description: Код приглашения
                      room:
                        $ref: '#/components/schemas/Room'
                      call:
                        $ref: '#/components/schemas/Call'
        '403':
          description: Нет доступа к комнате
        '404':
          description: Звонок не найден
        '500':
          description: Ошибка при генерации приглашения

  # Room Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Call:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID звонка
        room_id:
          type: string
          format: uuid
          description: ID комнаты
        initiator_id:
          type: string
          format: uuid
          description: ID инициатора звонка
        type:
          type: string
          enum: [audio, video]
          description: Тип звонка
        status:
          type: string
          enum: [initiated, connected, ended, rejected]
          description: Статус звонка
        start_time:
          type: string
          format: date-time
          description: Время начала звонка
        end_time:
          type: string
          format: date-time
          description: Время завершения звонка
        duration:
          type: integer
          description: Длительность звонка в секундах
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID комнаты
        name:
          type: string
          description: Название комнаты
        created_by:
          type: string
          format: uuid
          description: ID создателя комнаты
        invite_code:
          type: string
          description: Код приглашения в комнату
        is_active:
          type: boolean
          description: Активна ли комната
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SignalingMessage:
      type: object
      required:
        - type
        - payload
        - to
      properties:
        type:
          type: string
          enum: [offer, answer, ice-candidate]
          description: Тип сигнала WebRTC
        payload:
          type: object
          description: Данные сигнала
        from:
          type: string
          format: uuid
          description: ID отправителя сигнала
        to:
          type: string
          format: uuid
          description: ID получателя сигнала
        roomId:
          type: string
          format: uuid
          description: ID комнаты звонка 